// part_9/9.8-9.13_PATIENTOR/PATIENTOR_backend/src/types/types.ts
export interface Diagnose {
  code: string;
  name: string;
  latin?: string;
}

export interface Patient {
  id: string;
  name: string;
  dateOfBirth: string;
  ssn: string;
  gender: Gender;
  occupation: string;
}

export enum Gender {
  male = "male",
  female = "female",
  other = "other"

}


export type PatientSafe = Omit<Patient, "ssn">;
export type NewPatient = Omit<Patient, "id">;